+++ test/support/mocks/mongo-client-mock.ts
@@ -5,6 +5,7 @@ const mongoClientMock = (sandbox: SinonSandbox) => ({
   find: sandbox.stub(),
   toArray: sandbox.stub(),
   findOne: sandbox.stub(),
+  insertOne: sandbox.stub(),
   collection: sandbox.stub(),
   isConnected: sandbox.stub()
 })