+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Added [POST] `/contacts/` endpoint to create a new contact
+
+### Updated
+- Updated test suite for the `contacts-controller`, `contacts-service` and `contacts-repository` to account for the new
+functionality of being able to create a contact
+
 ## [0.3.0] - 2020-03-12
 ### Added
 - Added an MIT license for anyone to use the application